requirements.txt
django
python-dotenv
djangorestframework
django-ckeditor
